requirements.txt
fastapi==0.112.2
PyYAML==6.0.2
tensorflow==2.17.0
uvicorn==0.30.6